Update tests
Currently, we are working on implementing an API for simplifying library functionalities, which includes the CommonData, MavenData, MesonData, and CargoData structures. This implementation is still in progress and is aimed at facilitating testing without the need for a separate {config.toml} file to define parameters to be passed to templates. By doing so, we aim to streamline the testing process and make it more efficient.