Update binary
Two changes have been made to the code. Firstly, the 'project_path()' function has been removed from the library as it was not being used. Secondly, the previous functionality that was only available in the library has now also been added to the binary. This means that the 'cargo' command can now be run as 'cargo new', and you can use the optional '--lib' and '--ci' flags to request a library project or just CI files respectively. As part of these changes, the README.md file has been updated to provide explanations for the previous implementations.