Update mod.rs
The updated code now includes a modified version of the common_template() function, called common_template_skip(). This function allows you to skip some folders within the path passed during the testing phase. You can specify the folders you prefer to skip by highlighting them in the SKIPPED_FOLDER constant, which is passed as a parameter to the function.