cargo: Update testing process
The cargo-fuzz tool has been superseded by proptest within the cargo template. Consequently, the `cargo new` command can now create a tests folder that includes proptest.rs for use in the project.

As for integrating proptest in lib, further research is required to fully grasp its functionality. For the time being, it is considered a work in progress.